Add files via upload
Face mask detection is a computer vision technology that can detect whether a person is wearing a
mask or not in real-time video or image. By using pre-trained deep learning models and training them on a dataset of images,
this technology can accurately detect whether a person is wearing a mask or not.